+++ src/components/app/app.tsx
@@ -5,8 +5,7 @@ import FavoritesScreen from '../../pages/favorites-screen/favorites-screen.tsx';
 import NotFoundScreen from '../../pages/not-found-screen/not-found-screen.tsx';
 import OfferScreen from '../../pages/offer-screen/offer-screen.tsx';
 import PrivateRoute from '../private-route/private-route.tsx';
-import { AuthorizationStatus } from '../constants/status.tsx';
-import { AppRoute } from '../constants/app-route.tsx';
+import {AppRoute, AuthorizationStatus } from '../constants/constants.ts';
 
 
 type AppScreenProps = {
