--- src/components/constants/app-route.tsx
@@ -1,6 +1,0 @@
-export enum AppRoute {
-  Main = '/',
-  Login = '/login',
-  Favorites = '/favorites',
-  Offer = '/offer/:id'
-}
