--- src/components/constants/constants.tsx
@@ -1,3 +1,0 @@
-export const Setting = {
-  placeCount: 312
-};
