--- src/components/constants/status.tsx
@@ -1,6 +1,0 @@
-export enum AuthorizationStatus {
-  Auth = 'AUTH',
-  NoAuth = 'NO_AUTH',
-  Unknown = 'UNKNOWN',
-}
-
