+++ src/components/private-route/private-route.tsx
@@ -1,6 +1,6 @@
 import {Navigate} from 'react-router-dom';
-import { AuthorizationStatus } from '../constants/status';
 import { AppRoute } from '../constants/app-route';
+import { AuthorizationStatus } from '../constants/constants';
 
 type PrivateRouteProps = {
   authorizationStatus: AuthorizationStatus;
