+++ src/index.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
-import {Setting} from './components/constants/constants';
+import { Setting } from './components/constants/constants';
+
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
